+++ packages/backend-proxy-middleware/src/base/proxy.ts
@@ -21,7 +21,7 @@ import translations from './i18n.json';
 import type { ApiHubSettings, ApiHubSettingsKey, ApiHubSettingsService, BackendSystem } from '@sap-ux/store';
 import { AuthenticationType, BackendSystemKey, getService } from '@sap-ux/store';
 import { updateProxyEnv } from './config';
-import type { Url } from 'url';
+import type { Url } from 'node:url';
 import { addOptionsForEmbeddedBSP } from '../ext/bsp';
 import { getProxyForUrl } from 'proxy-from-env';
 import type { Socket } from 'node:net';