+++ packages/fiori-app-sub-generator/test/unit/utils/command-runner.test.ts
@@ -1,6 +1,6 @@
 import { CommandRunner } from '../../../src/utils/command-runner';
 import { initI18nFioriAppSubGenerator } from '../../../src/utils/i18n';
-import { platform } from 'os';
+import { platform } from 'node:os';
 import childProcess from 'child_process';
 // eslint-disable-next-line @typescript-eslint/no-var-requires
 const mockSpawn = require('mock-spawn');