+++ packages/ui5-proxy-middleware/src/base/utils.ts
@@ -1,4 +1,5 @@
 import type { ClientRequest, IncomingMessage, ServerResponse } from 'http';
+import type { Options } from 'http-proxy-middleware';
 import type { ToolsLogger } from '@sap-ux/logger';
 import { getMinimumUI5Version, type Manifest } from '@sap-ux/project-access';
 import type { RequestHandler, NextFunction, Request, Response } from 'express';
@@ -12,7 +13,7 @@ import {
     SANDBOX2_LINK,
     SANDBOX2_REPLACE_REGEX
 } from './constants';
-import type { Url } from 'url';
+import type { Url } from 'node:url';
 import { t } from '../i18n';
 import type { ReaderCollection } from '@ui5/fs';
 import type { Socket } from 'node:net';
@@ -345,4 +346,27 @@ export function directLoadProxy(
             next(error);
         }
     };
+}
+
+/**
+ * Create a rewrite based on the provided configuration.
+ * This will either replace the configured path with the provided pathReplace value or
+ * will prepend the UI5 version to the path.
+ * It also ensures that the path from the http-proxy-middleware starts with the
+ * config path in case of nested router instances.
+ *
+ * @param config proxy configuration
+ * @param ui5Ver UI5 version string
+ * @returns a path rewrite
+ */
+export function getPathRewrite(config: ProxyConfig, ui5Ver: string): Options['pathRewrite'] {
+    if (config.pathReplace) {
+        // Remove trailing slash from pathReplace if present
+        const sanitizedPathReplace = config.pathReplace?.replace(/\/$/, '');
+        return (path: string) =>
+            path.startsWith(config.path)
+                ? path.replace(config.path, sanitizedPathReplace)
+                : sanitizedPathReplace + path;
+    }
+    return (path: string) => (path.startsWith(config.path) ? ui5Ver + path : ui5Ver + config.path + path);
 }