+++ examples/ui-prompting-examples/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @sap-ux-private/ui-prompting-examples
 
+## 0.4.3
+
+### Patch Changes
+
+-   43a2446: chore: fix Sonar issues
+-   Updated dependencies [43a2446]
+    -   @sap-ux/ui-prompting@0.5.2
+
+## 0.4.2
+
+### Patch Changes
+
+-   Updated dependencies [7f652f3]
+    -   @sap-ux/ui-components@1.26.11
+    -   @sap-ux/ui-prompting@0.5.1
+
 ## 0.4.1
 
 ### Patch Changes