+++ packages/playwright/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @sap-ux-private/playwright
 
+## 0.2.1
+
+### Patch Changes
+
+-   43a2446: chore: fix Sonar issues
+
 ## 0.2.0
 
 ### Minor Changes